Rename MoE down-gather helper module
